leafwing_input_manager
Module axislike

Tools for working with directional axis-like user inputs (game sticks, D-Pads and emulated equivalents)

Structs

AxisConversionError

An AxisType could not be converted into a more specialized variant

DualAxis

Two directional axes combined as one input.

DualAxisData

A wrapped Vec2 that represents the combination of two input axes.

SingleAxis

A single directional axis with a configurable trigger zone.

VirtualAxis

A virtual Axis that you can get a value between -1 and 1 from.

VirtualDPad

A virtual DPad that you can get an DualAxis from.

Enums

AxisDirection

The directions for single-axis inputs.

AxisType

The type of axis used by a UserInput.

DualAxisDirection

The directions for dual-axis inputs.

DualAxisType

An axis for dual-axis inputs.

MouseMotionAxisType

The motion direction of the mouse.

MouseWheelAxisType

The motion direction of the mouse wheel.