Module leafwing_input_manager::dynamic_action

The dynamic action type.

This type is primarily intended for use when declaring every single action in a single enum isn’t feasible, for example due to a complex crate hierarchy where various feature crates want to individually declare their own actions without needing to coordinate with other crates.

Example:

// In crate `feature_one`
#[derive(DynActionMarker)]
struct FeatureOneAction;
struct FeatureOnePlugin;

impl Plugin for FeatureOnePlugin {
    fn build(&self, app: &mut App) {
        app.register_action::<FeatureOneAction>();
    }
}

// In crate `feature_two`
#[derive(DynActionMarker)]
struct FeatureTwoAction;
struct FeatureTwoPlugin;

impl Plugin for FeatureTwoPlugin {
    fn build(&self, app: &mut App) {
        app.register_action::<FeatureTwoAction>();
    }
}

// In crate `top_level_crate`, which depends on `feature_one` and `feature_two`
let mut app = App::new();
app.insert_resource(DynActionRegistry::get().unwrap())
    .add_plugins(FeatureOnePlugin)
    .add_plugins(FeatureTwoPlugin);
app.world.remove_resource::<DynActionRegistry>().unwrap().finish();

Structs

DynAction

The runtime representation of actions declared via marker types

DynActionRegistry

Coordinates the registration of dynamic action types

Traits

DynActionMarker

Trait implemented by marker types meant to be used as actions

RegisterActionToAppExt

Helper trait for registering DynAction types to an app where the DynActionRegistry exists as a resource

Derive Macros

DynActionMarker