Struct leafwing_input_manager::input_map::InputMap

pub struct InputMap<A: Actionlike> { /* private fields */ }

A Multi-Map that allows you to map actions to multiple UserInputs.

Many-to-One Mapping

You can associate multiple UserInputs (e.g., keyboard keys, mouse buttons, gamepad buttons) with a single action, simplifying handling complex input combinations for the same action. Duplicate associations are ignored.

One-to-Many Mapping

A single UserInput can be mapped to multiple actions simultaneously. This allows flexibility in defining alternative ways to trigger an action.

Clash Resolution

By default, the InputMap prioritizes larger UserInput combinations to trigger actions. This means if two actions share some inputs, and one action requires all the inputs of the other plus additional ones; only the larger combination will be registered.

This avoids unintended actions from being triggered by more specific input combinations. For example, pressing both S and Ctrl + S in your text editor app would only save your file (the larger combination), and not enter the letter s.

This behavior can be customized using the ClashStrategy resource.

Examples

use bevy::prelude::*;
use leafwing_input_manager::prelude::*;
use leafwing_input_manager::user_input::InputKind;

// Define your actions.
#[derive(Actionlike, Clone, Copy, PartialEq, Eq, Hash, Reflect)]
enum Action {
    Move,
    Run,
    Jump,
}

// Create an InputMap from an iterable,
// allowing for multiple input types per action.
let mut input_map = InputMap::new([
    // Multiple inputs can be bound to the same action.
    // Note that the type of your iterators must be homogeneous.
    (Action::Run, KeyCode::ShiftLeft),
    (Action::Run, KeyCode::ShiftRight),
    // Note that duplicate associations are ignored.
    (Action::Run, KeyCode::ShiftRight),
    (Action::Jump, KeyCode::Space),
])
// Associate actions with other input types.
.with(Action::Move, VirtualDPad::wasd())
.with(Action::Move, DualAxis::left_stick())
// Associate an action with multiple inputs at once.
.with_one_to_many(Action::Jump, [KeyCode::KeyJ, KeyCode::KeyU]);

// You can also use methods like a normal MultiMap.
input_map.insert(Action::Jump, KeyCode::KeyM);

// Remove all bindings to a specific action.
input_map.clear_action(&Action::Jump);

// Remove all bindings.
input_map.clear();

Implementations

impl<A: Actionlike> InputMap<A>

pub fn handle_clashes( &self, action_data: &mut HashMap<A, ActionData>, input_streams: &InputStreams<'_>, clash_strategy: ClashStrategy, )

Resolve clashing inputs, removing action presses that have been overruled

The usize stored in pressed_actions corresponds to Actionlike::index

impl<A: Actionlike> InputMap<A>

pub fn new( bindings: impl IntoIterator<Item = (A, impl Into<UserInput>)>, ) -> Self

Creates an InputMap from an iterator over action-input bindings. Note that the type of your iterators must be homogeneous.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn with(self, action: A, input: impl Into<UserInput>) -> Self

Associates an action with a specific input. Multiple inputs can be bound to the same action.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn with_one_to_many( self, action: A, inputs: impl IntoIterator<Item = impl Into<UserInput>>, ) -> Self

Associates an action with multiple inputs.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn with_multiple( self, bindings: impl IntoIterator<Item = (A, impl Into<UserInput>)>, ) -> Self

Adds multiple action-input bindings with the same input type.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

impl<A: Actionlike> InputMap<A>

pub fn insert(&mut self, action: A, input: impl Into<UserInput>) -> &mut Self

Inserts a binding between an action and a specific input. Multiple inputs can be bound to the same action.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn insert_one_to_many( &mut self, action: A, inputs: impl IntoIterator<Item = impl Into<UserInput>>, ) -> &mut Self

Inserts bindings between the same action and multiple UserInputs. Note that the type of your iterators must be homogeneous.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn insert_multiple( &mut self, bindings: impl IntoIterator<Item = (A, impl Into<UserInput>)>, ) -> &mut Self

Inserts multiple action-input bindings. Note that the type of your iterators must be homogeneous.

This method ensures idempotence, meaning that adding the same input for the same action multiple times will only result in a single binding being created.

pub fn insert_chord( &mut self, action: A, buttons: impl IntoIterator<Item = impl Into<InputKind>>, ) -> &mut Self

Insert a mapping between the simultaneous combination of buttons and the action provided

Any iterator convertible into a InputKind can be supplied, but will be converted into a HashSet for storage and use. Chords can also be added with the insert method if the UserInput::Chord variant is constructed explicitly.

When working with keyboard modifier keys, consider using the insert_modified method instead.

pub fn insert_modified( &mut self, action: A, modifier: Modifier, input: impl Into<InputKind>, ) -> &mut Self

Inserts a mapping between the simultaneous combination of the Modifier plus the input and the action provided.

When working with keyboard modifiers, should be preferred over insert_chord.

pub fn merge(&mut self, other: &InputMap<A>) -> &mut Self

Merges the provided InputMap into this map, combining their bindings, avoiding duplicates.

If the associated gamepads do not match, the association will be removed.

impl<A: Actionlike> InputMap<A>

pub const fn gamepad(&self) -> Option<Gamepad>

Fetches the Gamepad associated with the entity controlled by this input map.

If this is None, input from any connected gamepad will be used.

pub fn with_gamepad(self, gamepad: Gamepad) -> Self

Assigns a particular Gamepad to the entity controlled by this input map.

Use this when an InputMap should exclusively accept input from a particular gamepad.

If this is not called, input from any connected gamepad will be used. The first matching non-zero input will be accepted, as determined by gamepad registration order.

Because of this robust fallback behavior, this method can typically be ignored when writing single-player games.

pub fn set_gamepad(&mut self, gamepad: Gamepad) -> &mut Self

Assigns a particular Gamepad to the entity controlled by this input map.

Use this when an InputMap should exclusively accept input from a particular gamepad.

If this is not called, input from any connected gamepad will be used. The first matching non-zero input will be accepted, as determined by gamepad registration order.

Because of this robust fallback behavior, this method can typically be ignored when writing single-player games.

pub fn clear_gamepad(&mut self) -> &mut Self

Clears any Gamepad associated with the entity controlled by this input map.

impl<A: Actionlike> InputMap<A>

pub fn pressed( &self, action: &A, input_streams: &InputStreams<'_>, clash_strategy: ClashStrategy, ) -> bool

Checks if the action are currently pressed by any of the associated UserInputs.

Accounts for clashing inputs according to the ClashStrategy and remove conflicting actions.

pub fn process_actions( &self, input_streams: &InputStreams<'_>, clash_strategy: ClashStrategy, ) -> HashMap<A, ActionData>

Processes UserInput bindings for each action and generates corresponding ActionData.

Accounts for clashing inputs according to the ClashStrategy and remove conflicting actions.

impl<A: Actionlike> InputMap<A>

pub fn iter(&self) -> impl Iterator<Item = (&A, &Vec<UserInput>)>

Returns an iterator over all registered actions with their input bindings.

pub fn bindings(&self) -> impl Iterator<Item = (&A, &UserInput)>

Returns an iterator over all registered action-input bindings.

pub fn actions(&self) -> impl Iterator<Item = &A>

Returns an iterator over all registered actions.

pub fn get(&self, action: &A) -> Option<&Vec<UserInput>>

Returns a reference to the inputs associated with the given action.

pub fn get_mut(&mut self, action: &A) -> Option<&mut Vec<UserInput>>

Returns a mutable reference to the inputs mapped to action

pub fn len(&self) -> usize

Count the total number of registered input bindings.

pub fn is_empty(&self) -> bool

Returns true if the map contains no action-input bindings.

pub fn clear(&mut self)

Clears the map, removing all action-input bindings.

Keeps the allocated memory for reuse.

impl<A: Actionlike> InputMap<A>

pub fn clear_action(&mut self, action: &A)

Clears all input bindings associated with the action.

pub fn remove_at(&mut self, action: &A, index: usize) -> Option<UserInput>

Removes the input for the action at the provided index.

Returns Some(input) if found.

pub fn remove( &mut self, action: &A, input: impl Into<UserInput>, ) -> Option<usize>

Removes the input for the action if it exists

Returns Some with index if the input was found, or None if no matching input was found.