Trait leafwing_input_manager::input_processing::dual_axis::CustomDualAxisProcessor

pub trait CustomDualAxisProcessor: Send + Sync + Debug + DynClone + DynEq + DynHash + Reflect + Serialize {
    // Required method
    fn process(&self, input_value: Vec2) -> Vec2;
}

A trait for creating custom processor that handles dual-axis input values, accepting a Vec2 input and producing a Vec2 output.

Examples

use std::hash::{Hash, Hasher};
use bevy::prelude::*;
use bevy::utils::FloatOrd;
use serde::{Deserialize, Serialize};
use leafwing_input_manager::prelude::*;

/// Doubles the input, takes its absolute value,
/// and discards results that meet the specified condition on the X-axis.
// If your processor includes fields not implemented Eq and Hash,
// implementation is necessary as shown below.
// Otherwise, you can derive Eq and Hash directly.
#[derive(Debug, Clone, Copy, PartialEq, Reflect, Serialize, Deserialize)]
pub struct DoubleAbsoluteValueThenRejectX(pub f32);

// Add this attribute for ensuring proper serialization and deserialization.
#[serde_typetag]
impl CustomDualAxisProcessor for DoubleAbsoluteValueThenRejectX {
    fn process(&self, input_value: Vec2) -> Vec2 {
        // Implement the logic just like you would in a normal function.

        // You can use other processors within this function.
        let value = DualAxisSensitivity::all(2.0).scale(input_value);

        let value = value.abs();
        let new_x = if value.x == self.0 {
            0.0
        } else {
            value.x
        };
        Vec2::new(new_x, value.y)
    }
}

// Unfortunately, manual implementation is required due to the float field.
impl Eq for DoubleAbsoluteValueThenRejectX {}
impl Hash for DoubleAbsoluteValueThenRejectX {
    fn hash<H: Hasher>(&self, state: &mut H) {
        // Encapsulate the float field for hashing.
        FloatOrd(self.0).hash(state);
    }
}

// Remember to register your processor - it will ensure everything works smoothly!
let mut app = App::new();
app.register_dual_axis_processor::<DoubleAbsoluteValueThenRejectX>();

// Now you can use it!
let processor = DoubleAbsoluteValueThenRejectX(4.0);

// Rejected X!
assert_eq!(processor.process(Vec2::splat(2.0)), Vec2::new(0.0, 4.0));
assert_eq!(processor.process(Vec2::splat(-2.0)), Vec2::new(0.0, 4.0));

// Others are just doubled absolute value.
assert_eq!(processor.process(Vec2::splat(6.0)), Vec2::splat(12.0));
assert_eq!(processor.process(Vec2::splat(4.0)), Vec2::splat(8.0));
assert_eq!(processor.process(Vec2::splat(0.0)), Vec2::splat(0.0));
assert_eq!(processor.process(Vec2::splat(-4.0)), Vec2::splat(8.0));
assert_eq!(processor.process(Vec2::splat(-6.0)), Vec2::splat(12.0));

// The ways to create a DualAxisProcessor.
let dual_axis_processor = DualAxisProcessor::Custom(Box::new(processor));
assert_eq!(dual_axis_processor, DualAxisProcessor::from(processor));

Required Methods

fn process(&self, input_value: Vec2) -> Vec2

Computes the result by processing the input_value.