leafwing_input_manager::input_processing
Module single_axis

Processors for single-axis input values

Structs

AxisBounds

Specifies an acceptable min-max range for valid single-axis inputs, restricting all value stays within intended limits to avoid unexpected behavior caused by extreme inputs.

AxisDeadZone

A scaled version of AxisExclusion with the bounds set to AxisBounds::magnitude(1.0) that normalizes non-excluded input values into the “live zone”, the remaining range within the bounds after dead zone exclusion.

AxisExclusion

Specifies an exclusion range for excluding single-axis inputs, helping filter out minor fluctuations and unintended movements.

Enums

AxisProcessor

A processor for single-axis input values, accepting a f32 input and producing a f32 output.

Traits

CustomAxisProcessor

A trait for creating custom processor that handles single-axis input values, accepting a f32 input and producing a f32 output.

RegisterCustomAxisProcessorExt

A trait for registering a specific CustomAxisProcessor.

WithAxisProcessingPipelineExt

Provides methods for configuring and manipulating the processing pipeline for single-axis input.