Struct leafwing_input_manager::orientation::Direction

pub struct Direction { /* private fields */ }

A 2D unit vector that represents a direction

Its magnitude is always 1.0.

Example

use leafwing_input_manager::orientation::Direction;
use bevy::math::Vec2;

assert_eq!(Direction::NORTH.unit_vector(), Vec2::new(0.0, 1.0));
assert_eq!(Direction::try_from(Vec2::ONE), Ok(Direction::NORTHEAST));

assert_eq!(Direction::SOUTH * 3.0, Vec2::new(0.0, -3.0));
assert_eq!(Direction::EAST / 2.0, Vec2::new(0.5, 0.0));

Implementations

impl Direction

pub fn new(vec2: Vec2) -> Self

Creates a new Direction from a Vec2

The Vec2 stored internally will be normalized to have a magnitude of 1.0.

Panics

Panics if the length of the supplied vector has length zero or cannot be determined. Use try_from to get a Result instead.

pub const fn unit_vector(&self) -> Vec2

Returns the raw underlying Vec2 unit vector of this direction

This will always have a length of 1.0

impl Direction

pub const NORTH: Direction = _

The direction that points straight up

pub const EAST: Direction = _

The direction that points straight right

pub const SOUTH: Direction = _

The direction that points straight down

pub const WEST: Direction = _

The direction that points straight left

pub const NORTHEAST: Direction = _

The direction that points halfway between up and right

pub const SOUTHEAST: Direction = _

The direction that points halfway between down and right

pub const SOUTHWEST: Direction = _

The direction that points halfway between down and left

pub const NORTHWEST: Direction = _

The direction that points halfway between left and up