Trait leafwing_input_manager::orientation::Orientation

pub trait Orientation: Sized + Debug + From<Rotation> + Into<Rotation> + Copy {
    // Required method
    fn distance(&self, other: Self) -> Rotation;

    // Provided methods
    fn assert_approx_eq(self, other: impl Orientation) { ... }
    fn rotation_direction(&self, target: Self) -> RotationDirection { ... }
    fn rotate_towards(
        &mut self,
        target_orientation: Self,
        max_rotation: Option<Rotation>
    ) { ... }
}

A type that can represent an orientation in 2D space

Required Methods

fn distance(&self, other: Self) -> Rotation

Returns the absolute distance between self and other as a Rotation

The shortest path will always be taken, and so this value ranges between 0 and 180 degrees. Simply subtract the two rotations if you want a signed value instead.

Example

use leafwing_input_manager::orientation::{Orientation, Rotation};
use bevy::math::primitives::Direction2d;

Direction2d::Y.distance(Direction2d::from_xy(-1.0, -1.0).unwrap()).assert_approx_eq(Rotation::from_degrees(135.));

Provided Methods

fn assert_approx_eq(self, other: impl Orientation)

Asserts that self is approximately equal to other

Panics

Panics if the distance between self and other is greater than a hundredth of a degree.

fn rotation_direction(&self, target: Self) -> RotationDirection

Which RotationDirection is the shortest to rotate towards to reach target?

In the case of ties, RotationDirection::Clockwise will be returned.

Example

use leafwing_input_manager::orientation::{Orientation, RotationDirection};
use bevy::math::primitives::Direction2d;

assert_eq!(Direction2d::Y.rotation_direction(Direction2d::Y), RotationDirection::Clockwise);
assert_eq!(Direction2d::Y.rotation_direction(Direction2d::NEG_Y), RotationDirection::Clockwise);

assert_eq!(Direction2d::Y.rotation_direction(Direction2d::X), RotationDirection::Clockwise);
assert_eq!(Direction2d::Y.rotation_direction(Direction2d::NEG_X), RotationDirection::CounterClockwise);

assert_eq!(Direction2d::NEG_X.rotation_direction(Direction2d::NEG_Y), RotationDirection::CounterClockwise);
assert_eq!(Direction2d::NEG_Y.rotation_direction(Direction2d::NEG_X), RotationDirection::Clockwise);

fn rotate_towards( &mut self, target_orientation: Self, max_rotation: Option<Rotation> )

Rotates self towards target_orientation by up to max_rotation

Example

use leafwing_input_manager::orientation::{Rotation, Orientation};

let mut rotation = Rotation::SOUTH;

// Without a `max_rotation`, the orientation snaps
rotation.rotate_towards(Rotation::WEST, None);
assert_eq!(rotation, Rotation::WEST);

// With a `max_rotation`, we don't get all the way there
rotation.rotate_towards(Rotation::SOUTH, Some(Rotation::from_degrees_int(45)));
assert_eq!(rotation, Rotation::SOUTHWEST);

Object Safety

This trait is not object safe.