leafwing_input_manager
Module systems

The systems that power each InputManagerPlugin.

Functions

generate_action_diffs

Generates an Events stream of ActionDiff from ActionState

release_on_input_map_removed

Release all inputs when an InputMap<A> is removed to prevent them from being held forever.

swap_to_fixed_update

We are about to enter the FixedMain schedule, so we:

swap_to_update

We are about to enter the Main schedule, so we:

tick_action_state

Advances actions timer.

update_action_state

Fetches all the relevant ButtonInput resources to update ActionState according to the InputMap.

update_action_state_from_interaction

When a button with a component of type A is clicked, press the corresponding action in the ActionState