leafwing_input_manager::systems
Function update_action_state_from_interaction

pub fn update_action_state_from_interaction<A: Actionlike>(
    ui_query: Query<'_, '_, (&Interaction, &ActionStateDriver<A>)>,
    action_state_query: Query<'_, '_, &mut ActionState<A>>,
)

When a button with a component of type A is clicked, press the corresponding action in the ActionState

The action triggered is determined by the variant stored in your UI-defined button.