Trait leafwing_input_manager::typetag::Registry

pub trait Registry {
    type Identifier;
    type TraitObject: ?Sized;

    // Required methods
    fn register(
        &mut self,
        id: Self::Identifier,
        deserialize_fn: for<'de> fn(_: &mut dyn Deserializer<'de>) -> Result<Box<Self::TraitObject>, Error>,
    );
    fn get_deserialize_fn(
        &self,
        id: Self::Identifier,
    ) -> Result<&for<'de> fn(_: &mut dyn Deserializer<'de>) -> Result<Box<Self::TraitObject>, Error>, GetError<Self::Identifier>>;
    fn get_trait_object_name(&self) -> &'static str;

    // Provided methods
    fn register_type<T>(&mut self, id: Self::Identifier)
       where T: DeserializeOwned + Into<Box<Self::TraitObject>> { ... }
    fn deserialize_trait_object<'de, D>(
        &self,
        deserializer: D,
    ) -> Result<Box<Self::TraitObject>, <D as Deserializer<'de>>::Error>
       where D: Deserializer<'de>,
             Self: Sized,
             Self::Identifier: Deserialize<'de> + Debug { ... }
}

Registry mapping unique identifiers of types to their deserialize implementations, enabling deserialization of a specific trait object type.

Required Associated Types

type Identifier

The type of unique identifiers this registry uses. &'static str is used as a default.

type TraitObject: ?Sized

The trait object type this registry maps deserialize functions for.

Required Methods

fn register( &mut self, id: Self::Identifier, deserialize_fn: for<'de> fn(_: &mut dyn Deserializer<'de>) -> Result<Box<Self::TraitObject>, Error>, )

Register deserialize_fn as the deserialize function for id.

This method is infallible, but errors such as multiple registrations for id may be propagated to deserialization-time by making get_deserialize_fn return an error.

fn get_deserialize_fn( &self, id: Self::Identifier, ) -> Result<&for<'de> fn(_: &mut dyn Deserializer<'de>) -> Result<Box<Self::TraitObject>, Error>, GetError<Self::Identifier>>

Gets the deserialize function for id.

Errors

Implementations may return the following errors:

GetError::NotRegistered { id } if no deserialize function was registered for id.

GetError::MultipleRegistrations { id } if multiple deserialize functions were registered for id.

fn get_trait_object_name(&self) -> &'static str

Gets the trait object name, for diagnostic purposes.

Provided Methods

fn register_type<T>(&mut self, id: Self::Identifier)
where T: DeserializeOwned + Into<Box<Self::TraitObject>>,

Register a default deserialize function for type T as the deserialize function for id. T must implement DeserializeOwned and must be convertable into Box<Self::TraitObject> with Into<Box<Self::TraitObject>>.

This method is infallible, but errors such as multiple registrations for id may be propagated to deserialization-time by making get_deserialize_fn return an error.

fn deserialize_trait_object<'de, D>( &self, deserializer: D, ) -> Result<Box<Self::TraitObject>, <D as Deserializer<'de>>::Error>
where D: Deserializer<'de>, Self: Sized, Self::Identifier: Deserialize<'de> + Debug,

Deserialize a trait object with deserializer, using this registry to get the deserialize function for the concrete type, based on the deserialized ID.

Errors

Returns an error when get_deserialize_fn returns an error for the deserialized ID, or when deserialization fails.

Object Safety

This trait is not object safe.