leafwing_input_manager
Module user_input

Helpful abstractions over user inputs of all sorts

Structs

RawInputs

The basic input events that make up a UserInput.

Enums

InputKind

The different kinds of supported input bindings.

Modifier

A keyboard modifier that combines two KeyCode values into one representation.

UserInput

Some combination of user input, which may cross input-mode boundaries.