Trait leafwing_input_manager::input_mocking::MockInput

pub trait MockInput {
    // Required methods
    fn press_input(&mut self, input: impl Into<UserInput>);
    fn press_input_as_gamepad(
        &mut self,
        input: impl Into<UserInput>,
        gamepad: Option<Gamepad>,
    );
    fn send_axis_values(
        &mut self,
        input: impl Into<UserInput>,
        values: impl IntoIterator<Item = f32>,
    );
    fn send_axis_values_as_gamepad(
        &mut self,
        input: impl Into<UserInput>,
        values: impl IntoIterator<Item = f32>,
        gamepad: Option<Gamepad>,
    );
    fn release_input(&mut self, input: impl Into<UserInput>);
    fn release_input_as_gamepad(
        &mut self,
        input: impl Into<UserInput>,
        gamepad: Option<Gamepad>,
    );
    fn reset_inputs(&mut self);
}

Send fake input events for testing purposes

In game code, you should (almost) always be setting the ActionState directly instead.

Warning

You must call app.update() at least once after sending input with InputPlugin included in your plugin set for the raw input events to be processed into ButtonInput and Axis data.

Examples

use bevy::prelude::*;
use bevy::input::InputPlugin;
use leafwing_input_manager::axislike::MouseMotionAxisType;
use leafwing_input_manager::input_mocking::MockInput;
use leafwing_input_manager::prelude::*;

let mut app = App::new();

// This functionality requires Bevy's InputPlugin (included with DefaultPlugins)
app.add_plugins(InputPlugin);

// Press a key press directly.
app.press_input(KeyCode::KeyD);

// Or use chords to press multiple keys at the same time!
let bevy = [KeyCode::KeyB, KeyCode::KeyE, KeyCode::KeyV, KeyCode::KeyY];
app.press_input(UserInput::chord(bevy));

// Send values to an axis.
app.send_axis_values(SingleAxis::mouse_wheel_y(), [5.0]);

// Send values to two axes.
app.send_axis_values(DualAxis::mouse_motion(), [5.0, 8.0]);

// Release or deactivate an input.
app.release_input(KeyCode::KeyR);

// Reset all inputs to their default state.
app.reset_inputs();

// Remember to call the update method at least once after sending input.
app.update();

Required Methods

fn press_input(&mut self, input: impl Into<UserInput>)

Simulates an activated event for the given input, pressing all buttons and keys in the RawInputs of the input.

To avoid confusing adjustments, it is best to stick with straightforward button-like inputs, like KeyCodes, Modifiers, and UserInput::Chords. Axial inputs (e.g., analog thumb sticks) aren’t affected. Use Self::send_axis_values for those.

Input State Persistence

Pressed inputs remain active until explicitly released or by calling Self::reset_inputs.

Gamepad Input

Gamepad input is sent by the first registered controller. If no controllers are found, it is silently ignored.

Limitations

Unfortunately, due to upstream constraints, pressing a GamepadButtonType has no effect because Bevy currently disregards all external GamepadButtonChangedEvent events. See https://github.com/Leafwing-Studios/leafwing-input-manager/issues/516 for more details.

fn press_input_as_gamepad( &mut self, input: impl Into<UserInput>, gamepad: Option<Gamepad>, )

Simulates an activated event for the given input, using the specified gamepad, pressing all buttons and keys in the RawInputs of the input.

To avoid confusing adjustments, it is best to stick with straightforward button-like inputs, like KeyCodes, Modifiers, and UserInput::Chords. Axial inputs (e.g., analog thumb sticks) aren’t affected. Use Self::send_axis_values_as_gamepad for those.

Input State Persistence

Pressed inputs remain active until explicitly released or by calling Self::reset_inputs.

Limitations

Unfortunately, due to upstream constraints, pressing a GamepadButtonType has no effect because Bevy currently disregards all external GamepadButtonChangedEvent events. See https://github.com/Leafwing-Studios/leafwing-input-manager/issues/516 for more details.

fn send_axis_values( &mut self, input: impl Into<UserInput>, values: impl IntoIterator<Item = f32>, )

Simulates axis value changed events for the given input. Each value in the values iterator corresponds to an axis in the RawInputs of the input. Missing axis values default to 0.0.

To avoid confusing adjustments, it is best to stick with straightforward axis-like inputs like SingleAxis and DualAxis. Non-axial inputs (e.g., keys and buttons) aren’t affected; the current value will be retained for the next encountered axis. Use Self::press_input for those.

Input State Persistence

Each axis remains at the specified value until explicitly changed or by calling Self::reset_inputs.

Gamepad Input

Gamepad input is sent by the first registered controller. If no controllers are found, it is silently ignored.

fn send_axis_values_as_gamepad( &mut self, input: impl Into<UserInput>, values: impl IntoIterator<Item = f32>, gamepad: Option<Gamepad>, )

Simulates axis value changed events for the given input, using the specified gamepad. Each value in the values iterator corresponds to an axis in the RawInputs of the input. Missing axis values default to 0.0.

To avoid confusing adjustments, it is best to stick with straightforward axis-like inputs like SingleAxis and DualAxis. Non-axial inputs (e.g., keys and buttons) aren’t affected; the current value will be retained for the next encountered axis. Use Self::press_input_as_gamepad for those.

Input State Persistence

Each axis remains at the specified value until explicitly changed or by calling Self::reset_inputs.

fn release_input(&mut self, input: impl Into<UserInput>)

Simulates a released or deactivated event for the given input.

Gamepad Input

Gamepad input is sent by the first registered controller. If no controllers are found, it is silently ignored.

fn release_input_as_gamepad( &mut self, input: impl Into<UserInput>, gamepad: Option<Gamepad>, )

Simulates a released or deactivated event for the given input, using the specified gamepad.

fn reset_inputs(&mut self)

Resets all inputs in the MutableInputStreams to their default state.

All buttons are released, and just_pressed and just_released information on the ButtonInput type are lost. just_pressed and just_released on the ActionState will be kept.

This will clear all KeyCode, GamepadButton and MouseButton input streams, as well as any Interaction components and all input Events.

Object Safety

This trait is not object safe.