Trait leafwing_input_manager::input_mocking::QueryInput

pub trait QueryInput {
    // Required methods
    fn pressed(&self, input: impl Into<UserInput>) -> bool;
    fn pressed_on_gamepad(
        &self,
        input: impl Into<UserInput>,
        gamepad: Option<Gamepad>,
    ) -> bool;
    fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32>;
    fn read_axis_values_on_gamepad(
        &self,
        input: impl Into<UserInput>,
        gamepad: Option<Gamepad>,
    ) -> Vec<f32>;
}

Query input state directly for testing purposes.

In game code, you should (almost) always be using ActionState methods instead.

Examples

use bevy::prelude::*;
use bevy::input::InputPlugin;
use leafwing_input_manager::input_mocking::QueryInput;
use leafwing_input_manager::prelude::*;

let mut app = App::new();

// This functionality requires Bevy's InputPlugin (included with DefaultPlugins)
app.add_plugins(InputPlugin);

// Check if a key is currently pressed down.
let pressed = app.pressed(KeyCode::KeyB);

// Read the current vertical mouse scroll value.
let value = app.read_axis_values(SingleAxis::mouse_wheel_y());

// Read the current changes in relative mouse X and Y coordinates.
let values = app.read_axis_values(DualAxis::mouse_motion());
let x = values[0];
let y = values[1];

Required Methods

fn pressed(&self, input: impl Into<UserInput>) -> bool

Checks if the input is currently pressed or active.

This method is intended as a convenience for testing; use an InputMap in real code.

fn pressed_on_gamepad( &self, input: impl Into<UserInput>, gamepad: Option<Gamepad>, ) -> bool

Checks if the input is currently pressed or active on the specified Gamepad.

This method is intended as a convenience for testing; use an InputMap in real code.

fn read_axis_values(&self, input: impl Into<UserInput>) -> Vec<f32>

Retrieves the values on all axes represented by the input.

Binary inputs (e.g., keys and buttons) are treated like single-axis inputs, typically returning a value between 0.0 (not pressed) and 1.0 (fully pressed).

This method is intended as a convenience for testing; use an InputMap in real code.

fn read_axis_values_on_gamepad( &self, input: impl Into<UserInput>, gamepad: Option<Gamepad>, ) -> Vec<f32>

Retrieves the values on all axes represented by the input on the specified Gamepad.

Binary inputs (e.g., keys and buttons) are treated like single-axis inputs, typically returning a value between 0.0 (not pressed) and 1.0 (fully pressed).

This method is intended as a convenience for testing; use an InputMap in real code.

Object Safety

This trait is not object safe.