Module leafwing_input_manager::input_processing::dual_axis

Processors for dual-axis input values

Structs

CircleBounds

Specifies a circular region defining acceptable ranges for valid dual-axis inputs, with a radius defining the maximum threshold magnitude, restricting all values stay within intended limits to avoid unexpected behavior caused by extreme inputs.

CircleDeadZone

A scaled version of CircleExclusion with the bounds set to CircleBounds::new(1.0) that normalizes non-excluded input values into the “live zone”, the remaining range within the bounds after dead zone exclusion.

CircleExclusion

Specifies a cross-shaped region for excluding dual-axis inputs, with a radius defining the maximum excluded magnitude, helping filter out minor fluctuations and unintended movements.

DualAxisBounds

Specifies a square-shaped region defining acceptable ranges for valid dual-axis inputs, with independent min-max ranges for each axis, restricting all values stay within intended limits to avoid unexpected behavior caused by extreme inputs.

DualAxisDeadZone

A scaled version of DualAxisExclusion with the bounds set to DualAxisBounds::symmetric_all(1.0) that normalizes non-excluded input values into the “live zone”, the remaining range within the bounds after dead zone exclusion.

DualAxisExclusion

Specifies a cross-shaped region for excluding dual-axis inputs, with min-max independent min-max ranges for each axis, resulting in a per-axis “snapping” effect, helping filter out minor fluctuations to enhance control precision for pure axial motion.

DualAxisInverted

Flips the sign of dual-axis input values, resulting in a directional reversal of control.

DualAxisSensitivity

Scales dual-axis input values using a specified multiplier to fine-tune the responsiveness of control.

Enums

DualAxisProcessor

A processor for dual-axis input values, accepting a Vec2 input and producing a Vec2 output.

Traits

CustomDualAxisProcessor

A trait for creating custom processor that handles dual-axis input values, accepting a Vec2 input and producing a Vec2 output.

RegisterDualAxisProcessorExt

A trait for registering a specific CustomDualAxisProcessor.

WithDualAxisProcessingPipelineExt

Provides methods for configuring and manipulating the processing pipeline for dual-axis input.