Struct leafwing_input_manager::input_streams::MutableInputStreams

pub struct MutableInputStreams<'a> {

    pub gamepad_buttons: &'a mut Input<GamepadButton>,
    pub gamepad_button_axes: &'a mut Axis<GamepadButton>,
    pub gamepad_axes: &'a mut Axis<GamepadAxis>,
    pub gamepads: &'a mut Gamepads,
    pub gamepad_events: &'a mut Events<GamepadEvent>,
    pub keycodes: &'a mut Input<KeyCode>,
    pub scan_codes: &'a mut Input<ScanCode>,
    pub keyboard_events: &'a mut Events<KeyboardInput>,
    pub mouse_buttons: &'a mut Input<MouseButton>,
    pub mouse_button_events: &'a mut Events<MouseButtonInput>,
    pub mouse_wheel: &'a mut Events<MouseWheel>,
    pub mouse_motion: &'a mut Events<MouseMotion>,
    pub associated_gamepad: Option<Gamepad>,

}

A mutable collection of Input structs, which can be used for mocking user inputs.

These are typically collected via a system from the World as resources.

Fields
gamepad_buttons: &'a mut Input<GamepadButton>

A GamepadButton Input stream

gamepad_button_axes: &'a mut Axis<GamepadButton>

A GamepadButton Axis stream

gamepad_axes: &'a mut Axis<GamepadAxis>

A GamepadAxis Axis stream

gamepads: &'a mut Gamepads

A list of registered Gamepads

gamepad_events: &'a mut Events<GamepadEvent>

Events used for mocking gamepad-related inputs

keycodes: &'a mut Input<KeyCode>

A KeyCode Input stream

scan_codes: &'a mut Input<ScanCode>

A ScanCode Input stream

keyboard_events: &'a mut Events<KeyboardInput>

Events used for mocking keyboard-related inputs

mouse_buttons: &'a mut Input<MouseButton>

A MouseButton Input stream

mouse_button_events: &'a mut Events<MouseButtonInput>

Events used for mocking MouseButton inputs

mouse_wheel: &'a mut Events<MouseWheel>

A MouseWheel event stream

mouse_motion: &'a mut Events<MouseMotion>

A MouseMotion event stream

associated_gamepad: Option<Gamepad>

The Gamepad that this struct will detect inputs from

Implementations

impl<'a> MutableInputStreams<'a>

pub fn from_world(world: &'a mut World, gamepad: Option<Gamepad>) -> Self

Construct a MutableInputStreams from the World

pub fn guess_gamepad(&self) -> Option<Gamepad>

Guess which registered Gamepad should be used.

If an associated gamepad is set, use that. Otherwise use the first registered gamepad, if any.