Struct leafwing_input_manager::orientation::Rotation

pub struct Rotation { /* private fields */ }

A discretized 2-dimensional rotation

Internally, these are stored in millionths of a degree, and so can be cleanly added and reversed without accumulating error.

Example

use leafwing_input_manager::orientation::{Rotation, Direction, Orientation};
use core::f32::consts::{FRAC_PI_2, PI, TAU};

let east = Rotation::from_radians(0.0);
let north = Rotation::from_radians(FRAC_PI_2);
let west = Rotation::from_radians(PI);

Rotation::default().assert_approx_eq(Rotation::from_radians(0.0));
Rotation::default().assert_approx_eq(Rotation::from_radians(TAU));
Rotation::default().assert_approx_eq(500.0 * Rotation::from_radians(TAU));

(north + north).assert_approx_eq(west);
(west - east).assert_approx_eq(west);
(2.0 * north).assert_approx_eq(west);
(west / 2.0).assert_approx_eq(north);

north.assert_approx_eq(Rotation::NORTH);

Direction::from(west).assert_approx_eq(Direction::WEST);

Implementations

impl Rotation

pub const fn new(micro_degrees: u32) -> Rotation

Creates a new Rotation from a whole number of millionths of a degree

Measured clockwise from midnight.

pub const fn micro_degrees(&self) -> u32

Returns the exact internal measurement, stored in millionths of a degree

Measured clockwise from midnight (x=0, y=1). 360_000_000 make up a full circle.

impl Rotation

pub const DEGREE: u32 = 1_000_000u32

The number of micro-degrees in one degree

pub const HALF_CIRCLE: u32 = 180_000_000u32

The number of micro-degrees that make up a half circle

pub const FULL_CIRCLE: u32 = 360_000_000u32

The number of micro-degrees that make up a full circle

pub const NORTH: Rotation = _

The direction that points straight up

pub const EAST: Rotation = _

The direction that points straight right

pub const SOUTH: Rotation = _

The direction that points straight down

pub const WEST: Rotation = _

The direction that points straight left

pub const NORTHEAST: Rotation = _

The direction that points halfway between up and right

pub const SOUTHEAST: Rotation = _

The direction that points halfway between down and right

pub const SOUTHWEST: Rotation = _

The direction that points halfway between down and left

pub const NORTHWEST: Rotation = _

The direction that points halfway between left and up

impl Rotation

pub fn from_xy(xy: Vec2) -> Result<Rotation, NearlySingularConversion>

Constructs a Rotation from an (x,y) Euclidean coordinate

If both x and y are nearly 0 (the magnitude is less than EPSILON), [Err(NearlySingularConversion)] will be returned instead.

Example

use bevy::math::Vec2;
use leafwing_input_manager::orientation::Rotation;

assert_eq!(Rotation::from_xy(Vec2::new(0.0, 1.0)), Ok(Rotation::NORTH));

pub fn into_xy(self) -> Vec2

Converts this direction into an (x, y) pair with magnitude 1

pub fn from_radians(radians: impl Into<f32>) -> Rotation

Construct a Direction from radians, measured counterclockwise from the positive x axis

pub fn into_radians(self) -> f32

Converts this direction into radians, measured counterclockwise from the positive x axis

pub fn from_degrees(degrees: impl Into<f32>) -> Rotation

Construct a Direction from degrees, measured counterclockwise from the positive x axis

pub const fn from_degrees_int(degrees: u32) -> Rotation

Construct a Direction from a whole number of degrees, measured counterclockwise from the positive x axis

pub fn into_degrees(self) -> f32

Converts this direction into degrees, measured counterclockwise from the positive x axis