Module leafwing_input_manager::systems

The systems that power each InputManagerPlugin.

Functions

generate_action_diffs

Generates an Events stream of ActionDiff from ActionState

process_action_diffs

Generates an Events stream of ActionDiff from ActionState

release_on_disable

Release all inputs if the ToggleActions<A> resource exists and its enabled field is false.

release_on_input_map_removed

Release all inputs when an InputMap<A> is removed to prevent them from being held forever.

run_if_enabled

Uses the value of ToggleActions<A> to determine if input manager systems of type A should run.

tick_action_state

Advances actions timer.

update_action_state

Fetches all of the relevant Input resources to update ActionState according to the InputMap.

update_action_state_from_interaction

When a button with a component of type A is clicked, press the corresponding action in the ActionState